web_proj/requirements.txt
asgiref==3.3.1
autopep8==1.5.6
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Markdown==3.3.4
mysqlclient==2.0.3
pycodestyle==2.7.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
